backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pytest>=8.2,<9.0
httpx==0.25.2
openai==1.3.7
python-dotenv==1.0.0
pytest-asyncio==0.25.2
aiosqlite==0.20.0
email-validator==2.1.0.post1
requests==2.31.0
